+++ admin-panel/react-admin-interface/src/myLoginPage.js
@@ -10,7 +10,8 @@ const useStyles = makeStyles((theme) => ({
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center',
-  }}))
+  }
+}))
 
 
 const MyLoginPage = ({ theme }) => {
